+++ src/lasagna/stream.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import asyncio
 
 from typing import AsyncIterator, Iterable, TypeVar, Tuple, List, Union, Literal, cast
